fixing damage and melee attacks to use the roller, not the dice
changed health to an int, not int64, because that was dumb.

+++ combat/testcombatant.go
@@ -5,7 +5,7 @@ import "uuid"
 type TestCombatant struct {
 	id            uuid.UUID
 	name          string
-	curHealth     int64
+	curHealth     int
 	dead          bool
 	ac            int
 	resistance    map[DamageType]bool
@@ -38,7 +38,7 @@ func (t *TestCombatant) Name() string {
 	return t.name
 }
 
-func (t *TestCombatant) TakeMeleeDamage(damage int64) (isDead bool) {
+func (t *TestCombatant) TakeMeleeDamage(damage int) (isDead bool) {
 	t.curHealth = t.curHealth - damage
 	return t.curHealth <= 0
 }
